+++ CHANGELOG.md
@@ -6,6 +6,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file is similar to the format suggested by [Keep a CHANGELOG](https://github.com/olivierlacan/keep-a-changelog).
 
 ## Unreleased
+- [Patch] Update react-oui-icons version to 2.10.0 ([#1392](https://github.com/optimizely/oui/pull/1392))
+  - Update `icons.json` to use latest set of icons from v2.10.0
 
 ## 46.12.0 - 2020-07-31