+++ CHANGELOG.md
@@ -6,9 +6,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file is similar to the format suggested by [Keep a CHANGELOG](https://github.com/olivierlacan/keep-a-changelog).
 
 ## Unreleased
+
+- [Feature] Update Storybook to organize components ([#1381](https://github.com/optimizely/oui/pull/1381))
 - [Patch] Update react-oui-icons version to 2.10.0 ([#1392](https://github.com/optimizely/oui/pull/1392))
   - Update `icons.json` to use latest set of icons from v2.10.0
-
 - [Feature] Add `filterAvailableEntities` to **SearchPicker** for custom search-results filtering ([#1390](https://github.com/optimizely/oui/pull/1390)))
 
 ## 46.12.0 - 2020-07-31